コードをきれいにしている最中版(本来はこの書き方、体調不良の連鎖により焦りとバグ修正連鎖で修正祭りでコードが過去一汚くなってしまった。)

+++ 2025_01_制作_装甲騎兵ARES/プロジェクトファイル/ProjectCpp_2024/debugwall.cpp
@@ -116,10 +116,10 @@ void CDebugwall::InputpVtx()
     DATA EscData = GetDATA();
 
     //頂点座標の設定
-    pVtx[0].pos = D3DXVECTOR3((float)-PRINTSIZE, EscData.Pos.y, (float)PRINTSIZE);
-    pVtx[1].pos = D3DXVECTOR3((float)PRINTSIZE, EscData.Pos.y, (float)PRINTSIZE);
-    pVtx[2].pos = D3DXVECTOR3((float)-PRINTSIZE, EscData.Pos.y, (float)-PRINTSIZE);
-    pVtx[3].pos = D3DXVECTOR3((float)PRINTSIZE, EscData.Pos.y, (float)-PRINTSIZE);
+    pVtx[0].pos = D3DXVECTOR3(-PRINTSIZE, EscData.Pos.y, PRINTSIZE);
+    pVtx[1].pos = D3DXVECTOR3(PRINTSIZE, EscData.Pos.y, PRINTSIZE);
+    pVtx[2].pos = D3DXVECTOR3(-PRINTSIZE, EscData.Pos.y,-PRINTSIZE);
+    pVtx[3].pos = D3DXVECTOR3(PRINTSIZE, EscData.Pos.y, -PRINTSIZE);
 
     //法線ベクトルの設定
     pVtx[0].nor = D3DXVECTOR3(0.0f, 1.0f, 0.0f);
